requirements.txt
python-dotenv
pendulum